+++ Attributes/Commands.md
@@ -1,0 +1,13 @@
+# Attributes Documentation
+
+## CommandAttribute
+
+Marks a class or method as a command.<br>
+
+```csharp
+[Command]
+public void Command()
+{
+    // ...
+}
+```
