+++ Commands/Attributes/AdminRequiredAttribute.cs
@@ -1,3 +1,4 @@
+#if WINDOWS
 /// ======================================================================
 ///		CheetahToolbox: (https://github.com/CraigCraig/CheetahToolbox)
 ///				Project:  Craig's CheetahToolbox a Swiss Army Knife
@@ -6,8 +7,13 @@
 ///			Author: Craig Craig (https://github.com/CraigCraig)
 ///		License:     MIT License (http://opensource.org/licenses/MIT)
 /// ======================================================================
-namespace CheetahToolbox.Contexts;
+namespace CheetahToolbox.Commands;
 
-public class CommandContext
-{
-}
+using System;
+
+/// <summary>
+/// Attribute to mark a command as requiring admin privileges to be executed.
+/// </summary>
+[AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, Inherited = false)]
+public class AdminRequiredAttribute() : Attribute();
+#endif
