+++ Commands/ClearCommand.cs
@@ -8,9 +8,6 @@
 /// ======================================================================
 namespace CheetahToolbox.Commands;
 
-using Attributes;
-using Utils;
-
 [Command("Console", null, "Console Commands")]
 public class ConsoleCommands() : CommandBase()
 {
