+++ Commands/Commands.md
@@ -1,0 +1,20 @@
+# Command Documentation
+
+:warning: Command system is being reworked<br>
+
+Commands marked with a :spider: have bugs or issues<br>
+Commands marked with a :newspaper: need documentation.<br>
+Commands marked with a :warning: are not finished.<br>
+
+Commands work in CLI and in the console.<br>
+Flags work only in CLI.
+
+## CLI Flags
+
+- `-d` enbale debug mode :warning:
+- `-t` testing flag :warning:
+
+## Console Commands
+
+- `help` show help :spider:
+- `exit` :spider:
