+++ Commands/LsCommands.cs
@@ -6,14 +6,14 @@
 ///			Author: Craig Craig (https://github.com/CraigCraig)
 ///		License:     MIT License (http://opensource.org/licenses/MIT)
 /// ======================================================================
-#if WINDOWS
-namespace CheetahToolbox.Attributes;
+//namespace CheetahToolbox.Commands;
 
-using System;
-
-/// <summary>
-/// Attribute to mark a command as requiring admin privileges to be executed.
-/// </summary>
-[AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, Inherited = false)]
-public class AdminRequiredAttribute() : Attribute();
-#endif
+//[Command]
+//public class LsCommands() : CommandBase("ls", "")
+//{
+//    [Command]
+//    public void Default()
+//    {
+//        // TODO: Implement Changing Directories
+//    }
+//}
