+++ Information/Ideas.md
@@ -1,0 +1,15 @@
+# Ideas for Immediate Features
+
+
+
+## Ideas for Eventual Features
+
+### Package Management
+
+- [ ] Windows Service
+    - [ ] Update Chocolatey/Scoop/Winget Packages
+      - [ ] Only Run on Programs not Running
+
+### File Searching
+
+###
