+++ Managers/ModuleManager.cs
@@ -8,7 +8,6 @@
 /// ======================================================================
 namespace CheetahToolbox.Managers;
 
-using Contexts;
 using Modules;
 using System.Reflection;
