+++ Modules/ModuleBase.cs
@@ -8,8 +8,7 @@
 /// ======================================================================
 namespace CheetahToolbox.Modules;
 
-using Commands.Utils;
-using Contexts;
+using Commands;
 
 /// <summary>
 /// Base class for all modules.
